+++ src/aioverse/models/segments/__init__.py
@@ -1,10 +1,13 @@
 from .base_segment			import BaseSegment
 from .audio_segment			import AudioInputSegment
+from .audio_url_segment		import AudioUrlSegment
 from .file_segment			import FileSegment
 from .image_base64_segment	import ImageBase64Segment
 from .image_url_segment		import ImageUrlSegment
 from .text_segment			import TextSegment
 from .unknown_segment		import UnknownSegment
+from .video_input_segment	import VideoInputSegment
+from .video_url_segment		import VideoUrlSegment
 
 
 __all__ = [
@@ -13,10 +16,13 @@ __all__ = [
 	"BaseSegment",
 	
 	"AudioInputSegment",
+	"AudioUrlSegment",
 	"FileSegment",
 	"ImageBase64Segment",
 	"ImageUrlSegment",
 	"TextSegment",
-	"UnknownSegment"
+	"UnknownSegment",
+	"VideoInputSegment",
+	"VideoUrlSegment"
 	
 ]